refactor: restructure the dashboard into concern-based modules and reorganize templates + static assets
