+++ src/cli/dashboard/templates/mod.rs
@@ -1,0 +1,11 @@
+mod builders;
+mod views;
+
+pub use builders::{build_matrix, build_nested, build_variant_coverage};
+pub use views::{
+    AdrsTemplate, ArtifactDetailTemplate, ChromeTemplate, ConfigFile, DepLink, DeployedTemplate,
+    FileCard, FileCardGroup, FileGridTemplate, FilesTemplate, HarnessFiles, HarnessHooks,
+    HookDetailTemplate, HookEntry, HooksTemplate, IntegrityProblem, ModuleDetailTemplate,
+    ModulesTemplate, OverviewTemplate, ProvenanceTemplate, SchemaGroup, SearchPageTemplate,
+    SearchResultsTemplate, VariantsTemplate,
+};
