+++ templates/dashboard/base.html
@@ -5,11 +5,11 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>forge dashboard</title>
     <link rel="stylesheet" href="/static/dashboard.css">
-    <link rel="stylesheet" href="/static/highlight.css">
-    <script src="/static/htmx.min.js"></script>
-    <script src="/static/marked.min.js"></script>
-    <script src="/static/purify.min.js"></script>
-    <script src="/static/highlight.min.js"></script>
+    <link rel="stylesheet" href="/static/vendor/highlight.css">
+    <script src="/static/vendor/htmx.min.js"></script>
+    <script src="/static/vendor/marked.min.js"></script>
+    <script src="/static/vendor/purify.min.js"></script>
+    <script src="/static/vendor/highlight.min.js"></script>
     <script src="/static/annotate.js"></script>
     <script src="/static/diff.js"></script>
     <script src="/static/mdfootnotes.js"></script>
