+++ templates/dashboard/pages/artifact_detail.html
