+++ templates/dashboard/pages/hook_detail.html
