+++ templates/dashboard/pages/modules.html
@@ -36,7 +36,7 @@
     </div>
     <div id="module-detail">
         {% if let Some(module) = detail %}
-        {% include "dashboard/module_detail.html" %}
+        {% include "dashboard/pages/module_detail.html" %}
         {% else %}
         <p style="color:var(--text-muted)">Select a repository from the sidebar.</p>
         {% endif %}
