+++ templates/dashboard/pages/search_page.html
@@ -58,6 +58,6 @@
 </div>
 
 <div id="search-results">
-{% include "dashboard/search_results.html" %}
+{% include "dashboard/partials/search_results.html" %}
 </div>
 {% endblock %}
